requirements.txt
langchain
langchain-core
langchain-community

# LLM Providers

langchain-openai
langchain-mistralai
openai
mistralai
langchain
langchain-community
langchain-huggingface
langchain-chroma
langchain-mistralai

transformers
sentence-transformers
huggingface_hub

chromadb
pypdf

python-dotenv

# Vector Database

chromadb

# Embedding Models

sentence-transformers

# Document Loaders dependencies

pypdf
unstructured
beautifulsoup4
lxml

# Text Processing

tiktoken

# Environment Variables

python-dotenv

# Data Handling

pandas
numpy

# API / App Layer (optional but common in RAG apps)

fastapi
uvicorn

# UI (optional)

streamlit

# Utilities

tqdm
requests